Revert "refactor: switch to dynamic route segment instead of search param, but fully static"
This reverts commit 816c7bc42dc83e5ed079edc871e46bc8a637806a.
