--- apps/dashboard/src/app/(splash)/layout.tsx
@@ -1,13 +1,0 @@
-import { Splash } from '../../components/Splash'
-
-export default function Layout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return (
-    <Splash>
-      {children}
-    </Splash>
-  )
-}
