+++ apps/dashboard/src/app/layout.tsx
@@ -1,5 +1,6 @@
 import { Inter } from 'next/font/google'
 import { Toaster } from 'sonner'
+import { Splash } from '../components/Splash'
 import './globals.css'
 
 const inter = Inter({ subsets: ['latin'] })
@@ -12,7 +13,9 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={inter.className}>
-        {children}
+        <Splash>
+          {children}
+        </Splash>
         <Toaster closeButton richColors />
       </body>
     </html >
