+++ apps/dashboard/src/app/page.tsx
@@ -1,11 +1,11 @@
-import { HomeSplash } from "../../components/Splash/HomeSplash";
+import { HomeSplash } from "../components/Splash/HomeSplash";
 
 export const metadata = {
   title: 'OG Studio',
   description: 'Figma-like OG (Open Graph) Image builder.',
 }
 
-// export const dynamic = 'force-static'
+export const dynamic = 'force-static'
 
 export default function Home() {
   return <HomeSplash />
