+++ apps/dashboard/src/app/templates/page.tsx
@@ -1,11 +1,11 @@
-import { TemplatesSplash } from "../../../components/Splash/TemplatesSplash";
+import { TemplatesSplash } from "../../components/Splash/TemplatesSplash";
 
 export const metadata = {
   title: 'Templates - OG Studio',
   description: 'Pre-made Open Graph image templates.',
 }
 
-// export const dynamic = 'force-static'
+export const dynamic = 'force-static'
 
 export default function Templates() {
   return <TemplatesSplash />
